backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pydantic==1.10.9
python-dotenv==1.0.0
sqlalchemy==1.4.50
alembic==1.11.1
psycopg2-binary==2.9.7
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
stripe==5.0.0
pytest==7.4.0
pytest-asyncio==0.22.0
httpx==0.24.1
python-multipart==0.0.6